move to bid_adjustments.rs
