+++ crates/rbuilder/src/building/bid_adjustments.rs
@@ -1,0 +1,125 @@
+use alloy_eips::Encodable2718 as _;
+use alloy_primitives::{Address, B256};
+use rbuilder_primitives::mev_boost::{
+    ssz_roots::{tx_ssz_leaf_root, CompactSszTransactionTree},
+    BidAdjustmentStateProofs,
+};
+use reth::revm::database::StateProviderDatabase;
+use reth_primitives::TransactionSigned;
+use revm::{database::BundleAccount, Database as _};
+use std::collections::{hash_map, HashMap, HashSet};
+use tracing::*;
+
+use crate::building::{
+    cached_reads::CachedDB, BlockBuildingContext, BlockState, FinalizeError,
+    ThreadBlockBuildingContext, TransactionSszLeafRootCache,
+};
+
+/// Generate bid adjustment state proofs.
+pub fn generate_bid_adjustment_state_proofs(
+    block_state: &mut BlockState,
+    ctx: &BlockBuildingContext,
+    local_ctx: &mut ThreadBlockBuildingContext,
+) -> Result<HashMap<Address, BidAdjustmentStateProofs>, FinalizeError> {
+    if ctx.adjustment_fee_payers.is_empty() {
+        return Ok(Default::default());
+    }
+
+    let builder_signer = &ctx.builder_signer;
+    let builder_address = builder_signer.address;
+    let fee_recipient_address = ctx.attributes.suggested_fee_recipient;
+
+    let proof_targets = HashSet::from_iter(
+        [builder_address, fee_recipient_address]
+            .into_iter()
+            .chain(ctx.adjustment_fee_payers.clone()),
+    );
+
+    // Pre-load all proof targets that are missing from the bundle state.
+    // This is a requirement for accounts to become a part of the trie and be able to generate proofs for them.
+    let mut cachedb = CachedDB::new(
+        StateProviderDatabase::new(block_state.state_provider()),
+        &mut local_ctx.cached_reads,
+        &ctx.shared_cached_reads,
+    );
+    for fee_payer in &ctx.adjustment_fee_payers {
+        if let hash_map::Entry::Vacant(entry) =
+            block_state.bundle_state_mut().state.entry(*fee_payer)
+        {
+            let account_info = cachedb
+                .basic(*fee_payer)
+                .map_err(|error| FinalizeError::Other(error.into()))?;
+            entry.insert(BundleAccount {
+                original_info: account_info.clone(),
+                info: account_info,
+                status: revm::database::AccountStatus::Loaded,
+                storage: Default::default(),
+            });
+        }
+    }
+
+    let mut account_proofs =
+        ctx.root_hasher
+            .account_proofs(block_state.bundle_state(), &proof_targets, local_ctx)?;
+
+    let Some(builder_proof) = account_proofs.remove(&builder_address) else {
+        return Err(FinalizeError::Other(eyre::eyre!(
+            "account proof for builder {builder_address} is missing"
+        )));
+    };
+    let Some(fee_recipient_proof) = account_proofs.remove(&fee_recipient_address) else {
+        return Err(FinalizeError::Other(eyre::eyre!(
+            "account proof for proposer {fee_recipient_address} is missing"
+        )));
+    };
+
+    let mut bid_adjustments = HashMap::default();
+    for fee_payer_address in &ctx.adjustment_fee_payers {
+        let Some(fee_payer_proof) = account_proofs.remove(fee_payer_address) else {
+            error!(
+                %fee_payer_address,
+                "Fee payer proof is missing"
+            );
+            continue;
+        };
+
+        bid_adjustments.insert(
+            *fee_payer_address,
+            BidAdjustmentStateProofs {
+                builder_address,
+                builder_proof: builder_proof.clone(),
+                fee_recipient_address,
+                fee_recipient_proof: fee_recipient_proof.clone(),
+                fee_payer_address: *fee_payer_address,
+                fee_payer_proof,
+            },
+        );
+    }
+
+    Ok(bid_adjustments)
+}
+
+/// Compute the SSZ transaction proof for bid adjustments.
+pub fn compute_cl_placeholder_transaction_proof(
+    transactions: &[TransactionSigned],
+    cache: &TransactionSszLeafRootCache,
+) -> Vec<B256> {
+    let mut buf = Vec::new();
+    let mut leaves = Vec::with_capacity(transactions.len());
+    for tx in transactions {
+        let leaf = match cache.read().get(tx.hash()) {
+            Some(leaf) => *leaf,
+            None => {
+                buf.clear();
+                tx.encode_2718(&mut buf);
+                let leaf_root = tx_ssz_leaf_root(&buf);
+                cache.write().insert(*tx.hash(), leaf_root);
+                leaf_root
+            }
+        };
+        leaves.push(leaf);
+    }
+
+    let target = transactions.len().checked_sub(1).unwrap();
+    CompactSszTransactionTree::from_leaves(leaves).proof(target)
+}
